config/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2
beautifulsoup4==4.12.2
config==0.5.1
django==4.2.4; python_version >= '3.8'
django-bootstrap4==23.2
django-ckeditor==6.7.0
django-js-asset==2.1.0; python_version >= '3.8'
et-xmlfile==1.1.0
greenlet==2.0.2
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
pyodbc==4.0.39
python-dateutil==2.8.2
pytz==2023.3
pyxlsb==1.0.10
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.4.1; python_version >= '3.7'
sqlalchemy==2.0.20
sqlparse==0.4.4
typing-extensions==4.7.1; python_version >= '3.7'
tzdata==2023.3; python_version >= '2'
